+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Remove the `annotate-type` command (https://github.com/svenvanheugten/mutannot/pull/41)
+
 ## [0.9.0] - 2026-06-31
 
 ### Changed